+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## Next:
+## [1.2.0](https://github.com/Maarti/JenkinsDocExtension/releases/tag/1.2.0) - 2021-03-14
 
 - Documentation for Jenkins environment variables
 - Autocompletion for Jenkins environment variables